% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cy5.R
\docType{data}
\name{Cy5}
\alias{Cy5}
\title{Experimental data of freely diffusing Cy5 dye in water at a concentration of 33 nM.}
\format{
A matrix with 2048 columns and 5000 rows
}
\usage{
data(Cy5)
}
\description{
A single vector of single-point scan data, in the form of a TIFF file.
}
\details{
To analyze this data set with FCSlib, import by typing: readFileTiff("Cy5.tif")
}
\keyword{datasets}
