% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cy5_100nM.R
\docType{data}
\name{Cy5_100nM}
\alias{Cy5_100nM}
\title{Experimental data of free Cy5 molecules diffusing in water at a concentration of 100 nM.}
\format{
A data frame with 50000 rows and 2 variables
}
\usage{
data(Cy5_100nM)
}
\description{
A dataset containing a vector of a single-point, periodical acquisition.
}
\details{
\itemize{
  \item t is a vector Tau.
  \item f is a time point of aquisition.
}
}
\keyword{datasets}
