% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cy5_1nM.R
\docType{data}
\name{Cy5_1nM}
\alias{Cy5_1nM}
\title{Simulation of Cy5 diffusing in water at a concentration of 1nM.}
\format{A data frame with 50000 rows and 2 variables}
\usage{
data(Cy5_1nM)
}
\description{
A dataset containing a simulation vector of a single point acquisition. The variables are as follows:
}
\details{
\itemize{
  \item t. a vector Tau.
  \item f. the value of the acquisition over time.
}
}
\keyword{datasets}
