% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf.R
\name{pcf}
\alias{pcf}
\title{Pair Correlation Function}
\usage{
pcf(img, nPoints = 1000, type = "d", dr = 1)
}
\arguments{
\item{img}{The image to analyze}

\item{nPoints}{The size of the sub-vectors in which the input vectors will be divided. This number must be less than N/2.}

\item{type}{By defautl 'l'. The possible values are 'l' and 'c'}

\item{dr}{Is the distance between the two columns that will be correlated. For a value of deltar = 3, the columns are correlated as follows: (1,4), (2,5), ..., (n-3, n), with n as the last column.}
}
\value{
An image depicting the correlation between the pixel i and pixel i + dr.
}
\description{
The function pCF performs the correlation between the pixel i and pixel i + dr.
}
\details{
The pair correlation function (pCF) analyzes data of a periodically scanned line, measuring the time it takes a particle to go from one pixel to another, i.e. calculates the spatial cross-correlation function
between pixels.
G(tau,deltar) = (<F(t,0) F(t + tau, deltar)>/<F(t,0)> <F(t,deltar)>)-1
}
\examples{
dmv2 <- data.matrix(v2DataSet)
pB <- pcf(dmv2, nPoints = 2500, dr = 10, type = 'd')

\donttest{
# Pair Correlation Function Colum Distance with dr
dmv2 <- data.matrix(v2DataSet)
pB <- pcf(dmv2, nPoints = 5000, dr = 10, type = 'd')

# Plot the result
library("fields")
di <- dim(pB)
tau <- (1:(di[2]))
image.plot( x = 1:di[1], y = log10(tau), z = pB, main = "Column Distance 10",
xlab = "Pixel", ylab = "Logarithmic tau",
cex.lab = 1.2, cex.main = 1.2, cex.axis = 1)


# Pair Correlation Function Autocorrelacion
dmv2 <- data.matrix(v2DataSet)
pB <- pcf(dmv2, nPoints = 5000, type = 'a')

# Plot the result
library("fields")
di <- dim(pB)
tau <- (1:(di[2]))
image.plot( x = 1:di[1], y = log10(tau), z = pB, main = "Autocorrelation",
xlab = "Pixel", ylab = "Logarithmic tau",
cex.lab = 1.2, cex.main = 1.2, cex.axis = 1)


# Pair Correlation Function Fixed Colum with dr
dmv2 <- data.matrix(v2DataSet)
pB <- pcf(dmv2, nPoints = 5000, dr = 10, type = 'c')

# Plot the result
library("fields")
di <- dim(pB)
tau <- (1:(di[2]))
image.plot( x = 1:di[1], y = log10(tau), z = pB, main = "Fixed Colum 10",
xlab = "Pixel", ylab = "Logarithmic tau",
cex.lab = 1.2, cex.main = 1.2, cex.axis = 1)
}
}
\seealso{
\code{\link{fcs}, \link{pcomb}}
}
\author{
Raul Pinto Camara.
}
