% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileModel.R
\name{readFileModel}
\alias{readFileModel}
\title{Read File Model}
\usage{
readFileModel(filename)
}
\arguments{
\item{filename}{The name of the file to read from.}
}
\value{
params   A list containing the parameters as well as the model.
}
\description{
Reads a txt file and returns the parameters and the model (equation).
}
\details{
Read a txt file using the scan function and extracts the parameters and the model (equation) in the file.
}
\examples{
\donttest{
modelData <- readFileModel(filename)
}
}
\seealso{
\code{\link{fitFCS}}
}
\author{
Raúl Pinto Cámara.
}
